+++ src/add.js
@@ -1,5 +1,3 @@
-import removeTask from './remove.js';
-
 export class TaskObj {
   constructor(description, completed, index) {
     this.description = description;
@@ -15,6 +13,21 @@ export class TaskObj {
 
   static x = 0;
 }
+
+// remove and rearrange
+
+export const rearrange = (item, pos) => {
+  item.index = pos + 1;
+};
+
+const removeTask = (e) => {
+  const clickedRemove = e.target;
+  const parent = clickedRemove.parentNode;
+  parent.remove();
+  TaskObj.taskArr.splice(clickedRemove.id, 1);
+  TaskObj.taskArr.forEach(rearrange);
+  localStorage.setItem('Tasks', JSON.stringify(TaskObj.taskArr));
+};
 
 const editTask = (e) => {
   const clickedEdit = e.target;
@@ -97,4 +110,4 @@ const addItem = () => {
   }
 };
 
-export { addItem, editTask };
+export { addItem, editTask, removeTask };
